Time: 0 ms (100%), Space: 9.3 MB (63.08%) - LeetHub

+++ 3754-concatenate-non-zero-digits-and-multiply-by-sum-i/3754-concatenate-non-zero-digits-and-multiply-by-sum-i.c
@@ -1,5 +1,5 @@
 long long sumAndMultiply(int n) {
-    long long digit;
+    int digit;
     long long x= 0;
     long long place = 1;
     long long sum = 0;
